Update welcome document
I reviewed the Welcome to BU document and made some changes (grammatical and content-wise). I tried improving on some of the descriptions of things to make them more user-friendly, and slightly more descriptive.

Hopefully, this is a bit more in the right direction. If it's not, we can revert later.